run-mongod-replication: fix races.
This fixes MongoDB replication that was broken by
df104523417431ac013733b5478e483124b048c8 commit.

Before this change we run the replset_supervisor.sh script in the
background and after that run the mongodb daemon. Because
replset_supervisor.sh script connect to mongo it was possible that
database wasn't ready yet and script was failing. I updated the script
for waiting when mongo will accept connections.

Replica set started working but only for a short time and later it
stopped working again because mongo process was killed to early. To fix
this I removed the code from replset_supervisor.sh script that kills
mongo.