Remove call to mongo with no arguments
This hangs the tests forever if run with Docker 1.12 (tests work fine
with e.g., Docker 1.10).

The code is unnecessary now, as it seems to be a left-over of the
original way to check that SCL was enabled (introduced in #38). Now we
already have specific tests for SCL enablement, executed just two lines
above the line we're removing.